refactor(eidos): extract knowledge::causal submodule (#2642)
Move TemporalOrdering, CausalRelationType, CausalEdge to
knowledge/causal.rs. All items re-exported from mod.rs.
